+++ CHANGELOG.md
@@ -16,6 +16,23 @@ Those changes in added, changed or breaking changes, should include usage exampl
 * Added `eigen_common` dependency to the `eigensdk` crate when "full" feature is enabled in [#249](https://github.com/Layr-Labs/eigensdk-rs/pull/249).
 * Added bindings for `ECDSAStakeRegistry` and `ECDSAServiceManagerBase` in [#269](https://github.com/Layr-Labs/eigensdk-rs/pull/269).
 * Added release-plz in ci in [#275](https://github.com/Layr-Labs/eigensdk-rs/pull/275).
+* Added update_socket function for avs registry writer in [#268](https://github.com/Layr-Labs/eigensdk-rs/pull/268)
+  An example of use is the following:
+  ```rust
+  // Given an avs writer and a new socket address:
+
+  let tx_hash = avs_writer
+    .update_socket(new_socket_addr.into())
+    .await
+    .unwrap();
+
+  let tx_status = wait_transaction(&http_endpoint, tx_hash)
+    .await
+    .unwrap()
+    .status(); 
+  // tx_status should be true
+  ```
+
 * Added custom configuration for release-plz in [#281](https://github.com/Layr-Labs/eigensdk-rs/pull/281).
 
 ### Changed